修复在iPhone5s模拟器上不能通过编译的bug

+++ JokeClient-Swift/Utility/NSStringExt.swift
@@ -10,7 +10,7 @@ import Foundation
 
 extension String {
    
-    func stringHeightWith(fontSize:Float,width:Float)->Float
+    func stringHeightWith(fontSize:CGFloat,width:CGFloat)->CGFloat
     {
         var font = UIFont.systemFontOfSize(fontSize)
         var size = CGSizeMake(width,CGFLOAT_MAX)
